feat: adds fading to overlay for locations
adds two timers so that the highlight overlay is a nice and smooth
transition. modifies the timer so that there is some persistance and
adds a start and stop so that we can keep the two timers alive instead
of making them every time.

had to do this because otherwise you get multiple timers fighting
tweening the same thing and it would not behave as expected.