feat: adds map
adds the beginnings of the map, locations are defined in the cdb and
will show up on the map. the map is fixed to the graphic, can't be
changed.

the scenes will have a faded graphic of the map in the background with
props over it to make it look like it is at the location mentioned, but
without having to draw a new background for each one.