refactor: moves consts to Const, updates popup
moves most if not all const (static public / privates) into the Const
file so that everything is in one space.

updates popup so that it now uses bubble.text for parsing, so we can
have more fancy looking popups, really just want to be able to have the
item text different than the standard text.