feat: better time progression, implements visits take time
when you travel somewhere it takes time and will count against your
overall time, so you can only visit X amount of places, in our case 2.

also adds a disabling shader for the locations so when you run out of
time it shows the locations you can still go to (which are the locations
that you have already visited).