feat: adds InsertChoice condition, modifies forwarder
also makes the bubble text go on top when mouse overing.

modifies the behaviour of `Forwarder`. Originally i made it so that it
would take the valid choices and insert them into the current choices.
this is more like inserting than forwarding. This was ok for the one use
that i had before because there was only one choice, but i made another
one with multiple options and it wound up keeping some of the old
choices, so i renamed that to `InsertChoices` because that is more what
it does (merges choice wheels). `Forwarder` behavior is instead modified
to stop parsing choices and reset it, forwarding it to parse the new
choice set supplied.