assets: 90%s fox
the 1st full character, fox is pretty much finished. i only tested about
half of the dialogue. the other half is hard to get to because its is
more complex to collect the required items in order to test it. i need
to make a console or lower level access so i can set some things in
order to test it. maybe a text file it loads with some switches??

also creates a `Talked` and `NotTalked` that will give me some more
conditions when doing dialogue trees, but need to actually implement
this completely because logging isn't implemented yet.