assets: lamb 50%
implemented the sports storyline, tested about half of it.

also adds a debug console, but don't know how to place it correctly...
seems to be going all over the place.

this console lets me set and check variables, but i need to figure out
how to better place it. i think the s2d scaling is effect it adversly.