fix: text styling, empty actions
fixes text styling issues / allowing for different fonts by creating a
simple wrapped class that is a h2d.Text but contains what kind of font
data it has. i tried to compare the fonts, but this is difficult because
of how i am defining the fonts in Const. ideally i should make a new
font class that contains this information .. but don't have time for
that.

also modified the wheel to check if it has choices, and if it doesnt
then it aborts, this is so we can have a dialogue tree that ends and it
will kick you out of the conversation and you won't be able to talk to
that person anymore until things change.