feat: reworked the inventory to a backpack
changed the inventory because i need a way to be able to "use" items or
get more information about them? really i just want to be able to use
them but i'll need to add some more information on thing to give hits to
what they are used for and to teach people that they can click and
interact with items.