refactor: debug + fixes
refactors the debug drawing overlays, so i can toggle them in the game,
cleans up the display when trying to work through the dialogue and
understand the feel.

also fixes a few dialogue bugs with fox & book, and lamb & food

also fixes a blinking bug where the wheel would flicker when you start
outside of the bounds, because it would set opacity to 0.0 when you hit
that border.