feat: cleans up conversation choices
updates to the choice display for all choices so they are more obvious
what the text will be.

also makes the font smaller as well as adds a shadow and outline in
hopes to keep it readable while not making a crazy huge text display.

really need to bring these outside on their own layer...