feat: a tutorial quest & screen
adds a new screen replacing the `truestart` screen with something that
better explains what is going on in the game.

also makes a quick short quest from the weasel that makes you go through
it, though you can decide not to actually do it. its presented in a way
that everyone will be given the option and will see the colors, and
hopefully decided to do the quest. a short loop for them to see what the
game is about.